requirements.txt
google-cloud-compute
google-auth
